requirements.txt
# Core dependencies
pyyaml>=6.0

# Web interface
streamlit>=1.28.0
pandas>=2.0.0

# ML demonstration
scikit-learn>=1.3.0